Проект по 2-ому спринту, версия 2: 1. Изменено название класса Main. 2. Изменен уровень доступа к полю цвета яблок 3. Изменен вывод сум и добавлено форматирование.
